+++ src/literalizer/languages/elixir.py
@@ -44,6 +44,7 @@ from literalizer._formatters.format_integers import (
 from literalizer._formatters.format_strings import format_string_backslash_hash
 from literalizer._language import (
     NO_HETEROGENEOUS_BEHAVIOR,
+    NON_KEBAB_REF_CASES,
     CallStyle,
     CommentConfig,
     DateFormatConfig,
@@ -483,6 +484,9 @@ class Elixir(metaclass=LanguageCls):
         IdentifierCase.SNAKE,
         IdentifierCase.UPPER_SNAKE,
         IdentifierCase.PASCAL,
+    )
+    supported_ref_cases: ClassVar[frozenset[IdentifierCase]] = (
+        NON_KEBAB_REF_CASES
     )
 
     validate_spec_for_data = no_validate_spec_for_data