+++ src/literalizer/languages/objective_c.py
@@ -36,6 +36,7 @@ from literalizer._formatters.format_integers import (
 )
 from literalizer._language import (
     NO_HETEROGENEOUS_BEHAVIOR,
+    NON_KEBAB_REF_CASES,
     CommentConfig,
     DateFormatConfig,
     DatetimeFormatConfig,
@@ -514,6 +515,9 @@ class ObjectiveC(metaclass=LanguageCls):
     identifier_cases: ClassVar[tuple[IdentifierCase, ...]] = (
         IdentifierCase.CAMEL,
         IdentifierCase.PASCAL,
+    )
+    supported_ref_cases: ClassVar[frozenset[IdentifierCase]] = (
+        NON_KEBAB_REF_CASES
     )
 
     validate_spec_for_data = no_validate_spec_for_data